fix: add general error screen if permissions api call receives an error. Also refactor structure of common files
Closes #538
