+++ src/components/common/staticValues/commonLinks.ts
