+++ src/components/common/staticValues/types.ts
