+++ src/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
@@ -8,7 +8,8 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import { myMSALObj } from './auth/AuthConfig';
 import { getPermissions } from './services/Api';
 import { GeneralPermissions } from './components/common/interfaces';
-import NoApi from './components/common/NoApi';
+import NoApi from './components/common/informationalComponents/NoApi';
+import GeneralError from './components/common/informationalComponents/GeneralError';
 
 export const UserConfig = React.createContext(myMSALObj);
 export const Permissions = React.createContext<GeneralPermissions>({
@@ -22,6 +23,9 @@ export const Permissions = React.createContext<GeneralPermissions>({
 
 const renderApp = async (user) => {
     await getPermissions().then((result: any) => {
+        if (result && result.Message) {
+            return ReactDOM.render(<GeneralError />, document.getElementById('root'));
+        }
         if (result && result.admin !== undefined) {
             return ReactDOM.render(
                 <React.StrictMode>
